+++ lib/src/core/core.dart
@@ -21,4 +21,7 @@ export 'package:flutterguide/src/core/assets/assets.gen.dart';
 export 'widgets/image_type_enum.dart';
 
 export 'constants/constants.dart';
+export 'package:flutter_bloc/flutter_bloc.dart';
 
+export 'package:flutter_screenutil/flutter_screenutil.dart';
+export 'package:animated_tree_view/animated_tree_view.dart';
