+++ lib/src/core/widgets/scaffold_wrapper.dart
@@ -1,5 +1,4 @@
 import 'package:flutter/gestures.dart';
-import 'package:flutter/material.dart';
 
 import '../core.dart';
