+++ lib/src/features/dashboard/pages/basic_of_forex/basic_of_forex_page.dart
@@ -1,4 +1,3 @@
-import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'package:flutterguide/src/core/core.dart';
 
 class BasicOfForexPage extends StatelessWidget {
