+++ lib/src/features/dashboard/pages/basic_of_forex/pips_page.dart
@@ -1,0 +1,18 @@
+import 'package:flutterguide/src/core/core.dart';
+
+class PipsPage extends StatelessWidget {
+  const PipsPage({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return const ScaffoldWrapper(
+      body: CustomScrollView(
+        slivers: [
+          FlutterGuideAppBar(
+            title: "Learn Forex - Pips ",
+          )
+        ],
+      ),
+    );
+  }
+}
