+++ lib/src/features/dashboard/pages/introduction_page.dart
@@ -1,4 +1,3 @@
-import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'package:flutterguide/src/core/core.dart';
 
 class IntroductionPage extends StatelessWidget {
