+++ lib/src/features/dashboard/pages/pages.dart
@@ -1,1 +1,4 @@
 export 'introduction_page.dart';
+export 'basic_of_forex/basic_of_forex_page.dart';
+export 'basic_of_forex/pips_page.dart';
+export 'screen_view.dart';
